feat(core-wave1): extract calendar decode + week helper domains into taskify-core
