+++ taskify-core/src/index.ts
@@ -3,3 +3,6 @@ export * from "./events.js";
 export * from "./calendarPayload.js";
 export * from "./mutationContracts.js";
 export * from "./calendarProtocol.js";
+export * from "./calendarDecode.js";
+export * from "./weekDate.js";
+export * from "./weekRecurrence.js";
